Move the status bar and add the possibility to save when quiting
The status bar which allowed you to put in front the modelisation space window or the interview text window is now at the bottom of the application.
The zoom functionnality is also implemented in this bottom status bar.
When you quit without saving, the warning window allows you to save and quit.